move Repository type to common for eaiser testing of obligatron
